test(clickhouse): move URL/kwargs tests into the unit tree so CI runs them
tests/connectors/test_clickhouse.py is clickhouse-marked and the CI matrix
has no clickhouse row, so TestClickHouseUrlKwargs never ran in CI. The class
is pure kwargs assembly (no Docker), so it belongs in
tests/unit/test_clickhouse_helpers.py, which the unit job already executes.
Reuses the existing _FakeConnInfoFromUrl stand-in in place of _FakeChUrl.
